Remove non stl-compliant behaviour in half_precision Add more doc comments bit_writer and bit_reader

+++ include/bitstream/quantization/smallest_three.h
@@ -1,8 +1,5 @@
 #pragma once
 
-#include <cstdint>
-#include <cmath>
-
 /*
  *  Copyright (c) 2020 Stanislav Denisov, Maxim Munning, Davin Carten
  *
@@ -24,6 +21,9 @@
  *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  *  SOFTWARE.
  */
+
+#include <cstdint>
+#include <cmath>
 
 namespace bitstream
 {
@@ -134,7 +134,7 @@ namespace bitstream
 			float b = data.b * unpacker - half_range * unpacker;
 			float c = data.c * unpacker - half_range * unpacker;
 
-			float d = std::sqrt(1.0f - ((a * a) + (b * b) + (c * c)));
+			float d = std::sqrtf(1.0f - ((a * a) + (b * b) + (c * c)));
 
 			switch (data.m)
 			{
